Add formatOperatingSystem and move formatDate to utils/format.ts
