+++ client/src/components/PluginDetails/PluginMetadata.tsx
@@ -4,7 +4,7 @@ import { ReactNode } from 'react-markdown';
 import { Divider } from '@/components/common';
 import { MediaFragment } from '@/components/common/media';
 import { usePluginState } from '@/context/plugin';
-import { formatDate } from '@/utils/date';
+import { formatDate } from '@/utils';
 
 import { MetadataList } from './MetadataList';
 import { MetadataItem } from './PluginDetails.types';
